fix: Prevent import panic
