+++ py-polars/src/polars/_plr.pyi
@@ -2287,7 +2287,7 @@ def register_plugin_function(
     pass_name_to_apply: bool,
     changes_length: bool,
 ) -> PyExpr: ...
-def __register_startup_deps() -> None: ...
+def __register_startup_deps(warn_function: Callable[..., None]) -> None: ...
 def gen_uuid_v7() -> bytes: ...
 
 # functions.random
