+++ py-polars/src/polars/expr/list.py
@@ -9,8 +9,8 @@ from polars import exceptions
 from polars import functions as F
 from polars._utils.parse import parse_into_expression
 from polars._utils.unstable import unstable
-from polars._utils.various import issue_warning
 from polars._utils.wrap import wrap_expr
+from polars._warnings import issue_warning
 
 if TYPE_CHECKING:
     from collections.abc import Callable
